Day 0
Very simple scaffolding.
I'm a classic, single module maven with a fat-jar with the basics commons-io, commons-lang, junit and hamcrest
Taking fully advantage of the resource folders for the input files
And I went for reflection to instantiate the solution. Ugly but effective
What's going to bite me is rebuilding the jar everytime I want to re-run from a console, but the joys of using a semi-compiled language