Day 1
Arggg. I should read the problem.
I wasted hours because I thought I had to do a distinct. So I used sorted sets when it should have been a priority queue
I also wanted to use stream and reduces but in the end I stayed classic.

The second part was just building the histogram.

I added a nice scaffolding to hide the horrible regexp in java and do input + regexp -> stream of groups