Day 5: Block parsing and comparators
Another simple day. Build a list that contains the page and the pages that should be before. Reduce and maps for the win
For part2 I ended up using a simple sorting comparator after building much complex stuff that did not work

I took the opportunity to add the block text parsing and basic logging