Day 7: Fun with equations
A nice and fun day... I still remember how to solve equations
Once I figure out I needed to the recursion from right to left it was dead easy.
Part 2 did not work initially due to a bug, so I actually tried to do the classic queue solution calculating the value left to right... But while I was writing it, I spotted the bug.