Day 8: Geometry
It wasn't that difficult... but I struggled a lot to calculate the formulae
I assumed I needed lineal interpolation but I struggled to figure out how to apply it and also I wanted to figure out the formulae by my self rather than taking a quick refresher...
Then the dreaded moment of dealing with BigDecimal. I wish one day the do a wrapper similar to float/Float that instead of using Float uses BigDecimal... but in the meantime, I want
to do anything with decimals on Java and I don't want to suffer float/double lack of precision. So BigDecimal ugly syntax is the only way.

BigDecimal is safe and controlled but rounding and MathContext are complicated to figure out and I applied rounding to the lineal interpolation scale calculation instead of rounding the result points.
Once I figured that out the rest was kinda easy.

I also spend quite a lot of time building Space2D as Matrix is not good for geometry