Day 12: Path finding (part 1 only)
At the beginning I thought it was a simple matter of counting neighbours. And it was, but having the diagonals broke the region calculator :(
It took me hours to figure that out. In the mean, time I tried to build the fences polygon... I thought it was a waste of time, but it's going to be used on part 2
I need to work a little more on how to do the calculation for the polygon, but once I've got it, it should be easy