Day 13: BigDecimal is not that bad
I took the time to solve the equations. I'm so so rusty but once it was solved it was an easy implementation.
As I suspected part 2 implied big numbers and fortunately attempting remainder to see what was wrong give me the hint I needed more precision. I think I should do 64 by default
BigDecimal is ugly to understand but it's not that difficult to implement and it was really nice to work with in the end.
But yeah... you need to have a comment explaining every single operation and I more traditional way