Day 14: Visualise (not a proper solution)
Part 1 was easy, although the wrapping took me ages to have it right and because Position it's y,x while the examples are x,y there were a lot of bugs, I was preparing myself for looking for loops to calculate an insane number of cycles when the part 2 surprised me...
I went to reditt to see what a Christmas tree means when I saw that I just needed to do the visualization. I was playing with Day12 visualization (although I was writing to PNGs) so I add full visualization support (that have required a little bit of refactoring about the main runner)

If I have some spare time I'll code a solution. My idea is to try to look for rect lines
There are also changes on Day 12. Still working on it

PD: the visualization was done with Swing... It's actually not that bad to pain 2d on a canvas