Day 16 (part 1)
Another mess. I have struggled with something that I've done quite a lot of times before, priority queue + cache
In the end I managed to use Dijkstra. I tried it quite early but I discarded because I thought it was going to be impossible or
difficult to calculate the weights. In the end adding not a position, but a position+direction as node and adding as many nodes as entry points to a position did the trick

For part 2 this library does not give all the paths wih the best weight, just one. I'll see if having the weight allows me
to prune the priority queue/recursion in a reasonable time.