Day 18: Another day that Dijkstra is the MVP
So easy compared with last 2 days... glad that I spent time with JGraph on Day 16. With that and my matrix class this was incredible easy.
I think all the time spend on early days building the Matrix class had paid off by a lot. I don't think other years I managed to build something so reusable. Out of curiosity I've been comparing with my matrix last year, and it's much less versatile. Java it's not lean, it pushes you into building classes, interfaces and all the verbose stuff people complain about... But in the long term, it's so much easier to iterate or keep building the layers