Day 19: It should not have taken me that long
I was really stupid on how to tackle this kinda simple recursion problem...
I insisted in creating a has map with the sizes and iterating the pattern rather than taking slices and checking if that was a pattern 🤦
Then for part two I insisted on keeping a set with all the solutions. Which it should have been fine, but there are limits on what a set can hold and holding a number of permutations rather than the count itself was not efficient enoughs.

Moving to work with counts it was OK but I wasn't sure on how to cache it (an no, without cache it does not work). In the end it was not that difficult, it was a matter of breaking the actual patterns before (so the cache gets properly pre-filled)