Day 17: Part 2
This was a joined effort with evilteq
I discarded the reverse engineering effort and tried to accotate the ranges. I think eventually would have worked... But this is much better. Once I've got the hint that only the last 3 bits of the input affect the output it was easy to figure it out