Day 20 Part 1: Today it won't be Dijkstra's day
Part 2, I thought finding the possible shortcuts and then use Dijkstra was a good idea... It solves the problem, slowly (it takes a couple of minutes) but it won't work for part 2.
BFS + cache seems the best option as I could control the cheating state rather than calculating the shortcuts and calculating the best path over the modified map.