Day 23 I thought it was going to be worse...
It wasn't that difficult. I'm happier with my part 2 than my part 1
I used the JGraph so I don't have to build my own graph and it saved me quite a lot of time. For part 2 I went in descending order taking the combinations for a node's connections. Eg if I was testing group size 5, I grouped all the connections (including the source) in groups of five and tested for connectivity