Day 24 Part 1 only
Ok. Part 1 wasn't difficult. It took me far too long because I insisted on using JGrah when it was unnecessary.
For part 2, after looking at the logical gates to make a full binary addition https://www.101computing.net/binary-additions-using-logic-gates/ it's quite possible the input is a broken full binary addition so I've started to do checks to detect where the circuit fails. Then it should be easy to figure out the swaps