Day 14: Part 2
For once I decided to investigate a little bit before trying to brute force things...
Googling for how a binary addition is implemented via logic gates I found this https://www.101computing.net/binary-additions-using-logic-gates/
And upon closer inspection it was clear that indeed the input was clearly matching that.
After many iterations I managed to figure out a consistent way to get the full-adder for each digit and a test function to determine if a full-adder was right. With that I could swap wires (each full-adder is only 5, so it's reasonable), re-test until I found something that works