Day 16: Part 2
Finally... After trying a BFS and failing as it was impossible to do it and have all the paths y went back to Dijkstra. From each position in the given path with an alternative option, I run Dijkstra from it to find its best path. After some adjustments it wasn't that difficult in the end. I wasted far too many hours (again) on the BFS solution