Day 20: Part 2 The bane of my existence
After so many different attempts the problem is that I misread the example... arg arg arg!!!
My problem is that I was thinking I needed to count all the different paths from the cheat start until the cheat end.
My initial approach to recalculate Dijkstra after removing one vertex from each path point didn't scale well... So I build and BFS to solve part 1 with the intention to extend it for part 2 from 1 to 20... But also that did not scale well.
Then I had the idea about only needed to focus on the next points in the path from a cheat point within a 20 radius, but because I misread the problem I spend 2 days trying the impossible.

This solution in the end is slow as hell, but it runs in a reasonable time. Probably the BFS that I had a some point would have worked quite well with this, but I'm tired and I'll settle for this.

Only 21 part 2 to complete this year