Day 21: Part 2
In the end it wasn't that difficult... It took me ages to understand the key parts of the problem and why recursion/cache was the right solution.
My approach of translating the codes completely for each keypad it was not going to scale. Going move by move allowed me to avoid to get a 150Gb string in memory. I've got hinted about caching by the movement and the level

And that's a wrap. I did enjoy using Java, probably not the best tool for doing this, but it's not that bad and I rather have streams than python comprehensions.